pdca_api/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==4.2.15
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
idna==3.7
pillow==10.4.0
pycparser==2.22
PyJWT==1.7.1
PyMySQL==1.1.1
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
uWSGI==2.0.26
